minicrater: move all the implementation in driver.rs
